api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
beautifulsoup4
playwright
python-telegram-bot
tweepy
redis
anthropic
python-dotenv
apscheduler
pydantic
pydantic-settings
